Move scalarlist types to sa_types directory
